requirements.txt
numpy==2.4.0
opencv_python==4.12.0.88
Pillow==12.0.0
Requests==2.32.5
torch==2.9.1
git+https://github.com/facebookresearch/sam2.git